refactor: deepen four modules surfaced by the architecture review
Acting on the candidates from /improve-codebase-architecture. Four
refactors, each independently motivated but cohesive as a batch.

== 1. Collapse the easter-egg scattering (the big one) ==

FingerprintPlayground.vue was 724 lines with each egg smeared across 6
locations: type union, EGGS array, EGG_DURATIONS, eggBackground switch,
v-else-if template branch, and CSS block. Adding an egg meant editing
all 6.

Each egg is now a self-contained Vue SFC under
src/components/vue/fingerprint/eggs/. The playground imports a single
EGG_REGISTRY and renders the active egg via <component :is>. Adding an
egg is now: create one .vue file + append one entry to registry.ts.

  src/components/vue/FingerprintPlayground.vue   724 → 247 lines
  src/components/vue/fingerprint/types.ts        new (EggDef, EggProps)
  src/components/vue/fingerprint/registry.ts     new (15 entries)
  src/components/vue/fingerprint/eggs/*.vue      new (15 files)

Each egg owns its template, script (including any reactive state —
matrix tick, OTP XOR, Enigma rotors), style, and prefers-reduced-motion
rule. The engine stays tiny: hash, watch input, dispatch events, render
the active egg component.

== 2. Surface the implicit rnp:* event bus ==

Five custom events were referenced as string literals at ~12 sites
across SiteHeader, SiteSearch, FingerprintPlayground, EasterEggs, and
HeroDecrypt. No central registry; rename was unsafe.

src/lib/events.ts now exports RNP_EVENTS (and RNP_STORAGE for the
session key) as `as const` maps plus a dispatch() helper. All senders
and listeners import the same symbol; rename = one edit; typo =
compile error.

== 3. Deepen the vendor-source concept ==

RNP_VERSION was a hand-synced constant in content.config.ts that the
comment warned to "bump together with docs_ref". The man-page paths
were duplicated between fetch-sources.mjs (EXTRA_PATHS) and
content.config.ts (include regex).

src/lib/vendor-source.mjs now owns:
  - RNP_RELEASE_TAG (read from docs_ref in software/rnp.md)
  - RNP_VERSION (derived from the tag, stripping the leading 'v')
  - RNP_MAN_PAGES + RNP_EXTRA_SPARSE (shared sparse-checkout paths)

content.config.ts and fetch-sources.mjs both import. The hand-sync is
gone — bump docs_ref and RNP_VERSION follows.

src/lib/frontmatter.mjs extracts the YAML-front-matter parser shared
between asciidoc.ts (TS) and fetch-sources.mjs (Node). Kept in .mjs so
the standalone script can import it without a TS compiler.

== 4. Delete the duplicated helpers ==

decodeEntities was inlined verbatim inside renderAdoc (asciidoc.ts);
docSort was duplicated verbatim in two software route files. Both pass
the deletion test as pass-throughs.

  - renderAdoc now calls the existing decodeEntities function
  - docSort moves to src/lib/docs.ts; both routes import

== Skipped: front-matter parsing seam ==

The original review claimed fetch-sources.mjs and asciidoc.ts both
parse YAML front matter. Verifying on disk: fetch-sources.mjs doesn't
parse front matter itself — it sparse-clones repos and reads each
product's front matter via the shared parser now in frontmatter.mjs
(already done above as part of #3). The "seam" never existed as
described; nothing to do here.

== Verification ==

  - `npm run build` clean (no errors, no warnings)
  - HTTP 200 on /, /blog/, /software/rnp/, /advisories/
  - node -e import vendor-source.mjs → RNP_VERSION '0.18.1'
  - dev server HMR picked up all new files; egg files all 200
EOF
