+++ src/components/vue/fingerprint/eggs/RiboseEgg.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+defineProps<{ input: string; hex: string }>();
+</script>
+
+<template>
+  <div class="text-center font-sans leading-snug">
+    <p class="text-sm italic text-foreground">
+      "Open source is a development methodology; free software is a social movement."
+    </p>
+    <p class="mt-1.5 font-mono text-[0.7rem] text-faint">— Ribose · the people behind RNP</p>
+  </div>
+</template>
