+++ src/pages/software/[id].astro
@@ -5,20 +5,11 @@ import PageHero from '@/components/PageHero.astro';
 import Prose from '@/components/Prose.astro';
 import InstallTabs from '@/components/vue/InstallTabs.vue';
 import { tagLabel } from '@/lib/tags';
+import { docSort } from '@/lib/docs';
 
 export async function getStaticPaths() {
   const products = await getCollection('software');
   return products.map((entry) => ({ params: { id: entry.id }, props: { entry } }));
-}
-
-/** README first, then alphabetical by slug, with develop/* pages last. */
-function docSort(a: string, b: string): number {
-  if (a === 'README') return -1;
-  if (b === 'README') return 1;
-  const aDev = a.startsWith('develop/') ? 1 : 0;
-  const bDev = b.startsWith('develop/') ? 1 : 0;
-  if (aDev !== bDev) return aDev - bDev;
-  return a.localeCompare(b);
 }
 
 const { entry } = Astro.props;
