+++ src/pages/software/[product]/docs/[...slug].astro
@@ -1,6 +1,7 @@
 ---
 import { getCollection, render } from 'astro:content';
 import DocsLayout from '@/layouts/DocsLayout.astro';
+import { docSort } from '@/lib/docs';
 
 export async function getStaticPaths() {
   const entries = await getCollection('softwareDocs');
@@ -13,16 +14,6 @@ export async function getStaticPaths() {
       props: { entry },
     };
   });
-}
-
-/** README first, then alphabetical by slug, with develop/* pages last. */
-function docSort(a: string, b: string): number {
-  if (a === 'README') return -1;
-  if (b === 'README') return 1;
-  const aDev = a.startsWith('develop/') ? 1 : 0;
-  const bDev = b.startsWith('develop/') ? 1 : 0;
-  if (aDev !== bDev) return aDev - bDev;
-  return a.localeCompare(b);
 }
 
 const { entry } = Astro.props;
